+++ src/main/java/se/bjurr/gitchangelog/internal/settings/IssuesUtil.java
@@ -3,6 +3,7 @@ package se.bjurr.gitchangelog.internal.settings;
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static com.google.common.collect.Lists.newArrayList;
 import static se.bjurr.gitchangelog.internal.settings.SettingsIssueType.GITHUB;
+import static se.bjurr.gitchangelog.internal.settings.SettingsIssueType.GITLAB;
 import static se.bjurr.gitchangelog.internal.settings.SettingsIssueType.JIRA;
 
 import java.util.List;
@@ -18,6 +19,7 @@ public class IssuesUtil {
     List<SettingsIssue> issues = newArrayList(settings.getCustomIssues());
     addJira(issues);
     addGitHub(issues);
+    addGitLab(issues);
     return issues;
   }
 
@@ -26,6 +28,16 @@ public class IssuesUtil {
       if (settings.getGitHubApi().isPresent()) {
         issues.add(
             new SettingsIssue(GITHUB, "GitHub", settings.getGitHubIssuePattern(), null, null));
+      }
+    }
+  }
+
+  private void addGitLab(List<SettingsIssue> issues) {
+    if (settings.getGitLabIssuePattern().isPresent()) {
+      if (settings.getGitLabServer().isPresent()) {
+        issues.add(
+            new SettingsIssue(
+                GITLAB, "GitLab", settings.getGitLabIssuePattern().get(), null, null));
       }
     }
   }