+++ src/main/java/se/bjurr/gitchangelog/internal/settings/SettingsIssueType.java
@@ -3,5 +3,6 @@ package se.bjurr.gitchangelog.internal.settings;
 public enum SettingsIssueType {
   CUSTOM,
   JIRA,
-  GITHUB
+  GITHUB,
+  GITLAB
 }